Closed #8
Table view now renders tiles in order of groups and layers, but some
plugins/addons has no layer defined in their titles, so they will be
drawn in gray to test the coloring of the unimplemented tiles. As a
matter of viewing for now, the code pre-calculates the layout to ignore
the layers that has no plugin or addon.